+++ src/main/java/xyz/connect/user/web/controller/UserController.java
@@ -8,8 +8,8 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import xyz.connect.user.web.model.request.CreateUserRequest;
-import xyz.connect.user.web.model.request.LoginRequest;
+import xyz.connect.user.web.dto.request.CreateUserRequest;
+import xyz.connect.user.web.dto.request.LoginRequest;
 import xyz.connect.user.web.service.UserService;
 
 @RestController
