+++ src/main/resources/application-local.properties
@@ -1,1 +1,27 @@
-
+# H2 database configuration
+spring.datasource.url=jdbc:h2:tcp://localhost/~/connect_user1
+spring.datasource.username=sa
+spring.datasource.password=1234
+spring.datasource.driver-class-name=org.h2.Driver
+# Hibernate
+spring.jpa.show-sql=true
+spring.jpa.hibernate.ddl-auto=update
+spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.H2Dialect
+# jwt
+jwt.token.secret=cHJlIG9uIGNvbm5lY3Qgand0IHNlY3JleSBrZXkgcHJlIG9uIGNvbm5lY3Qgand0IHNlY3JleSBrZXkgcHJlIG9uIGNvbm5lY3Qgand0IHNlY3JleSBrZXkgcHJlIG9uIGNvbm5lY3Qgand0IHNlY3JleSBrZXkg=
+## multipart
+#spring.servlet.multipart.enabled=true
+#spring.servlet.multipart.max-file-size=100MB
+#spring.servlet.multipart.max-request-size=100MB
+## s3
+#cloud.aws.s3.bucket=grandchilds3
+#cloud.aws.stack.auto=false
+#cloud.aws.region.static=ap-northeast-2
+#cloud.aws.credentials.access-key=AKIATZ7IGGTM6R247B6Q
+#cloud.aws.credentials.secret-key=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
+#
+## Redis ?? ??
+#spring.redis.host=localhost
+#spring.redis.port=6379
+#spring.redis.password=
+#spring.redis.database=0
